fix: format number

+++ analytic_engine/src/instance/write.rs
@@ -126,7 +126,7 @@ pub enum Error {
         "Failed to write/flush table data, this table:{} does not belong to this worker: {}",
         source
     ))]
-    DataNotLegal { table: String, worker_id: u64 },
+    DataNotLegal { table: String, worker_id: usize },
 }
 
 define_result!(Error);
@@ -321,7 +321,7 @@ impl Instance {
         let worker_id = worker_local.worker_id();
         let worker_num = space.write_group.worker_num();
         ensure!(
-            table_data.id.as_u64() % worker_num != worker_id,
+            table_data.id.as_u64() as usize % worker_num != worker_id,
             DataNotLegal {
                 table: &table_data.name,
                 worker_id: worker_id
